+++ docs/RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## 0.2.0
+
+* Fully supported `parse()` with AST, tokens available
+* Fully supported `transpile()` with AST, tokens available
+
 ## 0.1.0
 
-* Partially support transpile
+* Partially supported `transpile()` with JS, TS, JSX, TSX, etc.